refactor: Refactor test and implementation classes in all modules to be present in the 'src' and 'test' directories, along standard Maven behavior; test whether tests are included as part of Maven test build step
NB: An error has been introduced with AudioLoaderImplTests.testLoadFromDirectory, but that is almost undoubtedly because it is badly-made. There is also an error loading a module properly for SoundImpl when running Maven tests, but I will delay investing time into this until (and only if) it appears after switching to Gradle.